build(matlab): add simulations for pernambuco
Add Matlab files, adapter for the Pernambuco scenario.

It is important to let it be noted that the simulations for Pernambuco
are in inital stage and incomplete. In order to properly model the
COVID-19 evolution in the state of Pernambuco, model parameters should
be updated.

Fitting model parameters for such a complex model is not trivial and
therefore, it was decided that this should be a future work.